[thy-types] basis
