+++ lib/smart_core/initializer/configuration.rb
@@ -25,7 +25,6 @@ module SmartCore::Initializer::Configuration
 
   # @since 0.1.0
   configuration do
-    # @since 0.1.0
     setting :default_type_system, :smart_types
     validate :default_type_system do |value|
       SmartCore::Initializer::TypeSystem.resolve(value) rescue false
