+++ lib/smart_core/initializer/errors.rb
@@ -39,6 +39,10 @@ module SmartCore::Initializer
 
   # @api public
   # @since 0.1.0
+  UnresolvedPluginDependencyError = Class.new(PluginError)
+
+  # @api public
+  # @since 0.1.0
   AlreadyRegisteredPluginError = Class.new(PluginError)
 
   # @api public
