+++ lib/smart_core/initializer/plugins/thy_types/thy_types.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module SmartCore::Initializer::TypeSystem
+  # @api public
+  # @since 0.1.0
+  class ThyTypes < Interop
+  end
+end
