+++ smart_initializer.gemspec
@@ -33,6 +33,7 @@ Gem::Specification.new do |spec|
   spec.add_dependency 'smart_engine', '~> 0.6'
   spec.add_dependency 'smart_types',  '~> 0.1.0'
   spec.add_dependency 'qonfig',       '~> 0.24'
+  spec.add_dependency 'thy',          '~> 0.1.4'
 
   spec.add_development_dependency 'bundler',          '~> 2.1'
   spec.add_development_dependency 'rake',             '~> 13.0'
